ai_testing_automation/requirements.txt
beautifulsoup4
requests
pandas
selenium
openpyxl
python-dotenv
langchain
langchain-google-genai